move championship in its own package
